added content count
